requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.108.0
google-auth==2.25.2
APScheduler==3.10.4
